Partial implementation of Basic WalkingAI

+++ CyberScroller/src/net/griscraft/cyberscroller/objects/GameObject.java
@@ -46,7 +46,7 @@ public abstract class GameObject {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 
@@ -54,7 +54,7 @@ public abstract class GameObject {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
